FIX : Abstraction for charts.
