+++ pages/tg/students/[slug].tsx
@@ -2,7 +2,6 @@ import TgsLayout from "components/Layouts/TgsLayout";
 import { prisma } from "lib/prisma";
 import {
   Student,
-  Subjects,
   FamilyDetails,
   Friends,
   Attendance,
@@ -15,22 +14,9 @@ import PersonalDetailForm from "components/DataForms/PersonalDetail";
 import FamilyDetailForm from "components/DataForms/FamilyDetail";
 import FriendsDetailForm from "components/DataForms/FriendsDetail";
 import AttendanceTable from "components/Tables/AttendanceTable";
-import {
-  Chart as ChartJS,
-  PointElement,
-  LineElement,
-  Filler,
-  Tooltip,
-  Legend,
-  CategoryScale,
-  LinearScale,
-  BarElement,
-} from "chart.js";
-import { Bar } from "react-chartjs-2";
-import { AssesmentType } from "@prisma/client";
-import ChartDataLabels from "chartjs-plugin-datalabels";
 import { useState } from "react";
 import { checkUserRoleAndRedirect } from "lib/checks";
+import AssesmentGraphs from "components/Graphs/AssesmentGraphs";
 
 export async function getServerSideProps(context: any) {
   const { params } = context;
@@ -57,18 +43,6 @@ export async function getServerSideProps(context: any) {
   });
 }
 
-ChartJS.register(
-  PointElement,
-  LineElement,
-  Filler,
-  Tooltip,
-  Legend,
-  CategoryScale,
-  LinearScale,
-  BarElement,
-  ChartDataLabels
-);
-
 const SingleStudentPage = ({
   student,
   familyDetails,
@@ -79,164 +53,6 @@ const SingleStudentPage = ({
   const [personalView, setPersonalView] = useState(true);
   const [statsView, setStatsView] = useState(false);
   const [goalsView, setGoalsView] = useState(false);
-
-  let data = {
-    labels: [Subjects.AI, Subjects.CN, Subjects.DP, Subjects.FE, Subjects.SEPM],
-    datasets: [],
-    options: {
-      animation: {
-        duration: 0,
-      },
-      plugins: {
-        legend: {
-          labels: {
-            // This more specific font property overrides the global property
-            font: {
-              size: 500,
-              weight: "bold",
-            },
-          },
-        },
-      },
-    },
-  };
-
-  let TAEdata = {
-    labels: [Subjects.AI, Subjects.CN, Subjects.DP, Subjects.FE, Subjects.SEPM],
-    datasets: [],
-    options: {
-      animation: {
-        duration: 0,
-      },
-      plugins: {
-        legend: {
-          labels: {
-            // This more specific font property overrides the global property
-            font: {
-              size: 500,
-              weight: "bold",
-            },
-          },
-        },
-      },
-    },
-  };
-
-  assesments.map((assesment) =>
-    TAEdata.datasets.push(
-      //@ts-ignore
-      assesment.name == AssesmentType.TAE
-        ? {
-            //@ts-ignore
-            label: assesment.name,
-            //@ts-ignore
-            data: [
-              assesment.score1,
-              assesment.score2,
-              assesment.score3,
-              assesment.score4,
-              assesment.score5,
-            ],
-            backgroundColor: "rgba(128, 0, 128, 0.5)",
-            pointHoverBackgroundColor: "#fff",
-            borderColor: "rgba(128, 0, 128, 0.3)",
-            datalabels: {
-              color: "gray",
-              anchor: "end",
-              align: "end",
-            },
-          }
-        : {}
-    )
-  );
-
-  assesments.map((assesment) =>
-    data.datasets.push(
-      //@ts-ignore
-      assesment.name == AssesmentType.CAE1
-        ? {
-            //@ts-ignore
-            label: assesment.name,
-            //@ts-ignore
-            data: [
-              assesment.score1,
-              assesment.score2,
-              assesment.score3,
-              assesment.score4,
-              assesment.score5,
-            ],
-            backgroundColor: "rgba(128, 0, 128, 0.5)",
-            pointHoverBackgroundColor: "#fff",
-            borderColor: "rgba(128, 0, 128, 0.3)",
-            datalabels: {
-              color: "gray",
-              anchor: "end",
-              align: "end",
-            },
-          }
-        : assesment.name == AssesmentType.CAE2
-        ? {
-            //@ts-ignore
-            label: assesment.name,
-            //@ts-ignore
-            data: [
-              assesment.score1,
-              assesment.score2,
-              assesment.score3,
-              assesment.score4,
-              assesment.score5,
-            ],
-            backgroundColor: "rgba(128, 0, 254, 0.5)",
-            pointHoverBackgroundColor: "#fff",
-            borderColor: "rgba(128, 0, 128, 0.3)",
-            datalabels: {
-              color: "gray",
-              anchor: "end",
-              align: "end",
-            },
-          }
-        : {}
-    )
-  );
-
-  const overallData = {
-    labels: [
-      "Sem 1",
-      "Sem 2",
-      "Sem 3",
-      "Sem 4",
-      "Sem 5",
-      "Sem 6",
-      "Sem 7",
-      "Sem 8",
-    ],
-    datasets: [
-      {
-        label: "Overall",
-        data: [
-          student.sem1Score,
-          student.sem2Score,
-          student.sem3Score,
-          student.sem4Score,
-          student.sem5Score,
-          student.sem6Score,
-          student.sem7Score,
-          student.sem8Score,
-        ],
-        backgroundColor: "rgba(0, 0, 255, 0.5)",
-        datalabels: {
-          color: "gray",
-          anchor: "end",
-          align: "end",
-        },
-      },
-    ],
-    options: {
-      animation: {
-        duration: 0,
-      },
-    },
-  };
 
   return (
     <TgsLayout>
@@ -353,30 +169,7 @@ const SingleStudentPage = ({
 
           {/* Graph's */}
           <div className={`flex flex-col ${statsView ? "" : "hidden"}`}>
-            <h1 className="text-2xl font-semibold mb-4">Stats : </h1>
-            <div className="flex">
-              <div className="w-[30rem] mr-5">
-                <h1 className="text-xl font-bold">
-                  Overall assesment Stats :{" "}
-                </h1>
-                <Bar
-                  //@ts-ignore
-                  data={overallData}
-                />
-              </div>
-              <div className="w-[30rem] mr-5">
-                <h1 className="text-xl font-bold">
-                  Continous Assesment Stats :{" "}
-                </h1>
-                <Bar data={data} />
-              </div>
-              <div className="w-[30rem] mr-5">
-                <h1 className="text-xl font-bold">
-                  Teacher Assesment Stats :{" "}
-                </h1>
-                <Bar data={TAEdata} />
-              </div>
-            </div>
+            <AssesmentGraphs assesments={assesments} student={student} />
           </div>
 
           {/* Goals */}
